move runpsscript
